requirements.txt
fastapi
uvicorn
streamlit
requests
pydantic
python-dotenv
langchain-groq
pydantic-settings
langchain
pdfplumber
python-docx
PyPDF2